+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.4.0 (2019-09-26)
+* Fixed multiple plugin binding namespacing
+* Added dump handling to plugins
+* Added manual class alias loader on binding instantiation
+
 ## v0.3.0 (2019-09-24)
 * Moved autoload code to Listener structure
 * Created singleton Register for default Listener